[chore]: #79 신고 디렉토리 정리
