+++ app/src/main/java/com/example/findu/presentation/ui/report/MissingReportFragment.kt
@@ -1,4 +1,4 @@
-package com.example.findu.presentation.ui.report.missing
+package com.example.findu.presentation.ui.report
 
 import android.Manifest
 import android.content.Intent
@@ -35,16 +35,15 @@ import androidx.navigation.fragment.findNavController
 import com.example.findu.databinding.FragmentMissingReportBinding
 import com.example.findu.presentation.type.PermissionType
 import com.example.findu.presentation.type.report.ReportType
-import com.example.findu.presentation.ui.report.constants.ReportConstants.IMAGE_RESULT_KEY
-import com.example.findu.presentation.ui.report.constants.ReportConstants.IMAGE_URI
+import com.example.findu.presentation.ui.report.constants.ReportConstants
 import com.example.findu.presentation.ui.report.dialog.ReportFinishDialog
 import com.example.findu.presentation.ui.report.dialog.ReportLocationActivity
-import com.example.findu.presentation.ui.report.dialog.ReportLocationDialog.Companion.POST_TAG
-import com.example.findu.presentation.ui.report.missing.navigation.MissingReportNavHost
-import com.example.findu.presentation.ui.report.missing.navigation.MissingReportRoute
-import com.example.findu.presentation.ui.report.missing.viewmodel.MissingReportUiEffect
-import com.example.findu.presentation.ui.report.missing.viewmodel.MissingReportUiEvent
-import com.example.findu.presentation.ui.report.missing.viewmodel.MissingReportViewModel
+import com.example.findu.presentation.ui.report.dialog.ReportLocationDialog
+import com.example.findu.presentation.ui.report.navigation.MissingReportNavHost
+import com.example.findu.presentation.ui.report.navigation.MissingReportRoute
+import com.example.findu.presentation.ui.report.viewmodel.MissingReportUiEffect
+import com.example.findu.presentation.ui.report.viewmodel.MissingReportUiEvent
+import com.example.findu.presentation.ui.report.viewmodel.MissingReportViewModel
 import com.google.accompanist.permissions.ExperimentalPermissionsApi
 import com.google.accompanist.permissions.isGranted
 import com.google.accompanist.permissions.rememberPermissionState
@@ -193,8 +192,8 @@ class MissingReportFragment : Fragment() {
     }
 
     private fun getCapturedUri() {
-        setFragmentResultListener(IMAGE_URI) { _, result ->
-            val imageUri = result.getString(IMAGE_RESULT_KEY)
+        setFragmentResultListener(ReportConstants.IMAGE_URI) { _, result ->
+            val imageUri = result.getString(ReportConstants.IMAGE_RESULT_KEY)
             imageUri?.let {
                 viewModel.handleEvent(MissingReportUiEvent.OnImageSelected(it.toUri()))
             }
@@ -206,7 +205,7 @@ class MissingReportFragment : Fragment() {
         resultLauncher =
             registerForActivityResult(ActivityResultContracts.StartActivityForResult()) { result ->
                 if (result.resultCode == AppCompatActivity.RESULT_OK) {
-                    val data = result.data?.getStringExtra(POST_TAG)
+                    val data = result.data?.getStringExtra(ReportLocationDialog.Companion.POST_TAG)
                     viewModel.handleEvent(MissingReportUiEvent.OnAddressUpdated(data ?: "주소 찾기 실패"))
                 }
             }
